backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pydantic-settings==2.5.2
python-dotenv==1.0.1
azure-identity==1.23.0
requests==2.32.3
